Replace ORDER BY by INTERNAL SORT BY in geomcache query
See
https://github.com/ad-freiburg/qlever/commit/aee40945a10a66aa24b1748b1652b4d51c2eec0e

Since that commit, the cache query that asks for all geometries took
very long for oms-planet because it uses ORDER BY, which has since
become a slow operation. For the purpose of the caching, INTERNAL SORT
BY is just as good, but faster (and also pinned to the cache for our
QLever instance).