Towards customizable SPARQL queries for the geom cache
The SPARQL queries for the `GeomCache` depend on the backend. We
currently have different queries for OSM and for Wikidata. So far, it
was hard-coded in several places, which of the queries was chosen. This
decition now happens in two dedicated methods `getQuery()` and
`getQueryCount()`. The mechanism is used to use the Wikidata queries
also for dblp-plus (which so far did not work).

TODO: The MapView for https://qlever.cs.uni-freiburg.de/dblp-plus/WF9LE4
now indeed works. But when you click on one of the points, the info box
is not shown, but instead "Session has been removed from cache". Why?