More efficient SPARQL query for counting the number of geometries
The previous query used `COUNT(DISTINCT ...)`, which amounts to a `GROUP
BY`, which is rather slow and crashed the latest version of OSM Planet.
Now just use the "all distinct geometries" query (which we have to
compute anyway), as subquery of a simple `COUNT`.

See also https://github.com/ad-freiburg/qlever/issues/1082