Try it with generic COUNT queries
So far, there was an explicit COUNT query for each group of backends.
Now the COUNT query is simply a variant of the query that gets the
geometries, namely with `SELECT (COUNT(?geometry) AS ?count) { ... }`
around it.

The old code is still there. Will clean this up in the next commit.