Fetch only non-point WKT geometries
So far, the initial cache-warming query fetched *all* `geo:asWKT`
geometries from the SPARQL endpoint. However, `POINT` geometries do
not need to be cached since (for QLever) the coordinates are folded
into the ID. In particular, this makes a huge difference for OSM Planet,
where the vast majority of geometries are `POINT` geometries (at the
time of this writing, over 20 billion out of 22 billion).